Moving to new react-navigation
