+++ client/BalanceApp/components/main/main.js
@@ -1,48 +1,41 @@
 // Vendors
 import React, { Component, PropTypes } from 'react';
-import { TouchableHighlight, Text } from 'react-native';
+import { TouchableHighlight, Text, Button } from 'react-native';
 import { connect } from 'react-redux';
 
 // Components
 import ProjectListContainer from '../project-list/project-list-container';
-import { styles } from './navigation-styles';
-import { fetchProjects } from '../../actions';
+import { styles } from '../navigation/navigation-styles';
 
 class MainScene extends Component {
   static propTypes = {
-    navigator: PropTypes.object.isRequired,
+    navigation: PropTypes.object.isRequired,
     dispatch: PropTypes.func.isRequired
+  }
+
+  static navigationOptions = {
+    title: () => <Text style={[styles.text, styles.title, styles.mainTitle]}>BALANCE</Text>,
+    header: ({ state, setParams }) => ({
+      style: { backgroundColor: '#333' },
+      left: (
+        <TouchableHighlight>
+          <Text style={[styles.button, styles.text]}>!?</Text>
+        </TouchableHighlight>
+      ),
+      right: (
+        <TouchableHighlight>
+          <Text style={[styles.button, styles.text]}>✚</Text>
+        </TouchableHighlight>
+      )
+    })
   }
 
   constructor (props) {
     super(props);
   }
 
-  projectDetailLeftButton (pop) {
-    function onPress () {
-      this.props.dispatch(fetchProjects());
-      pop();
-    }
-    return (
-      <TouchableHighlight onPress={onPress.bind(this)}>
-        <Text style={[styles.button, styles.text, { fontWeight: 'normal' } ]}>Back</Text>
-      </TouchableHighlight>
-    );
-  }
-
-  projectDetailTitle () {
-    return (<Text style={ [styles.title, styles.text, { letterSpacing: 1 }] }>Details</Text>)
-  }
-
   onProjectTap (project) {
-    this.props.navigator.push({
-      title: project.title,
-      scene: 'project-detail',
-      leftButton: this.projectDetailLeftButton(this.props.navigator.pop),
-      rightButton: () => null,
-      renderTitle: this.projectDetailTitle(),
-      passProps: { projectId: project._id }
-    });
+    this.props.navigation.navigate('Project', { projectId: project._id });
   }
 
   render () {
