--- client/BalanceApp/components/main/navigation-styles.js
@@ -1,20 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create  ({
-  view: {
-    backgroundColor: "#333333"
-  },
-  title: {
-    fontFamily: "Helvetica Neue",
-    letterSpacing: 5
-  },
-  button: {
-    paddingHorizontal: 20
-  },
-  text: {
-    paddingTop: 10,
-    fontSize: 18,
-    fontWeight: 'bold',
-    color: "#FFFFFF"
-  }
-});
