+++ client/BalanceApp/package.json
@@ -10,6 +10,7 @@
     "react": "15.4.2",
     "react-native": "0.40.0",
     "react-native-autogrow-textinput": "^3.0.1",
+    "react-navigation": "git+https://git@github.com/react-community/react-navigation.git",
     "react-redux": "^5.0.1",
     "redux": "^3.6.0",
     "redux-thunk": "^2.1.0"
